requirements.txt
fastapi[standard]
uvicorn[standard]==0.22.0
python-dotenv==1.1.1
supabase==2.18.0
httpx<0.29
pydantic==2.11.7
SQLAlchemy==2.0.22
asyncpg==0.27.0
